requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
openai==1.3.0
gpt4all==2.5.0
watchdog==3.0.0
pyyaml==6.0.1
sqlite-utils==3.35
sqlalchemy==2.0.23
redis==5.0.1
cryptography==41.0.7
python-dotenv==1.0.0
marshmallow==3.20.1
gunicorn==21.2.0
celery==5.3.4
prometheus-client==0.18.0
structlog==23.2.0

# Security testing dependencies
python-owasp-zap-v2.4>=0.9.0
requests>=2.31.0
python-jose>=3.3.0
safety>=2.3.0
bandit>=1.7.5
